Add GameCraft VAE (AutoencoderKLCausal3D) for Hunyuan-GameCraft port
- Add GameCraftVAE and EncoderCausal3D/DecoderCausal3D in gamecraftvae.py
- Add gamecraftvae_blocks.py with DownEncoderBlockCausal3D, UpDecoderBlockCausal3D,
  and GameCraftVAEAttention (softmax upcast to float32 for bf16/fp16 stability)
- Add GameCraftVAEConfig with scaling_factor, temporal_compression_ratio=4,
  mid_block_causal_attn=True, and block_out_channels layout
- Register AutoencoderKLCausal3D -> GameCraftVAE in models registry for checkpoint loading

+++ fastvideo/models/vaes/gamecraftvae_blocks.py
@@ -191,7 +191,8 @@ class GameCraftVAEAttention(nn.Module):
         attn = torch.matmul(q, k.transpose(-2, -1)) * scale
         if attention_mask is not None:
             attn = attn + attention_mask
-        attn = F.softmax(attn, dim=-1)
+        # Official uses upcast_softmax=True: compute softmax in fp32 for numerical stability
+        attn = F.softmax(attn.float(), dim=-1).to(q.dtype)
         hidden_states = torch.matmul(attn, v)
         hidden_states = hidden_states.transpose(1, 2).reshape(batch_size, seq_len, -1)
         hidden_states = self.to_out(hidden_states) + residual
